Removed old checkout controller
