+++ ShopClient/Controllers/Cart/CartViewController.swift
@@ -62,7 +62,7 @@ class CartViewController: BaseViewController<CartViewModel>, CartTableDataSource
     
     // MARK: - actions
     @IBAction func checkoutTapped(_ sender: UIButton) {
-        showCheckoutNewController()
+        showCheckoutController()
     }
     
     // MARK: - CartTableDataSourceProtocol
