+++ ShopClient/Extensions/UIStoryboardAdditions.swift
@@ -70,10 +70,6 @@ extension UIStoryboard {
         return UIStoryboard(name: "Checkout", bundle: nil)
     }
     
-    class func checkoutNew() -> UIStoryboard {
-        return UIStoryboard(name: "CheckoutNew", bundle: nil)
-    }
-    
     class func address() -> UIStoryboard {
         return UIStoryboard(name: "Address", bundle: nil)
     }
