--- ShopClient/Storyboards/CheckoutNew.storyboard
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13527"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Checkout New View Controller-->
-        <scene sceneID="Csf-kA-Wwy">
-            <objects>
-                <viewController storyboardIdentifier="CheckoutNewControllerIdentifier" id="MDv-XD-ZMR" customClass="CheckoutNewViewController" customModule="ShopClient" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="iHH-Hu-qii">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="grouped" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="1" estimatedSectionFooterHeight="1" translatesAutoresizingMaskIntoConstraints="NO" id="pyA-b5-EGg">
-                                <rect key="frame" x="0.0" y="20" width="375" height="647"/>
-                                <color key="backgroundColor" red="0.96078431372549022" green="0.96078431372549022" blue="0.96078431372549022" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                            </tableView>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="pyA-b5-EGg" firstAttribute="top" secondItem="4Jr-sw-EUI" secondAttribute="top" id="C11-H0-J03"/>
-                            <constraint firstItem="pyA-b5-EGg" firstAttribute="bottom" secondItem="4Jr-sw-EUI" secondAttribute="bottom" id="Wrs-B6-K3H"/>
-                            <constraint firstItem="pyA-b5-EGg" firstAttribute="leading" secondItem="4Jr-sw-EUI" secondAttribute="leading" id="kxC-Hd-SRK"/>
-                            <constraint firstItem="pyA-b5-EGg" firstAttribute="trailing" secondItem="4Jr-sw-EUI" secondAttribute="trailing" id="rlC-ne-3RB"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="4Jr-sw-EUI"/>
-                    </view>
-                    <connections>
-                        <outlet property="tableView" destination="pyA-b5-EGg" id="4nf-Ch-xLB"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="aWy-vV-BzK" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-50.399999999999999" y="133.5832083958021"/>
-        </scene>
-    </scenes>
-</document>
