+++ ShopClient/SupportingFiles/Localizable.strings
@@ -28,9 +28,6 @@
 "Button.Ok" = "OK";
 "Button.Cancel" = "Cancel";
 "Button.Checkout" = "Checkout";
-"Button.PaymentMethodWebsite" = "Pay on our website";
-"Button.PaymentMethodCard" = "Pay by card";
-"Button.PaymentMethodApplePay" = "Apple Pay";
 "Button.PaymentAddressViewContinue" = "Continue to shipping method";
 "Button.Continue" = "Continue";
 "Button.SeeAll" = "See All";
@@ -53,10 +50,6 @@
 "Label.PriceFrom" = "From %@ %@";
 "Label.Quantity" = "QTY:";
 "Label.PriceEach" = "%@ %@ each";
-"Label.SelectPaymentMethod" = "Please select a Payment Method";
-"Label.PaymentMethodWebsite" = "Use the card to pay for on our website";
-"Label.PaymentMethodCard" = "Use the card to pay for on our app";
-"Label.PaymentMethodApplePay" = "You can use any credit or debit card in you AppleID account (including your cards in Apple Pay) for quick, easy checkout";
 "Label.PriceOfProducts" = "Price of products: %.2f %@";
 "Label.PriceOfDelivery" = "Price of delivery: %.2f %@";
 "Label.TotalTax" = "Total tax: %.2f %@";
